# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "行空 API, 用于小说搜索",
        "title": "行空 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://xingkong.io/support",
            "email": "zhenhua32@xingkong.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/book/{id}": {
            "get": {
                "description": "返回小说的详情",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "返回小说的详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                }
            }
        },
        "/book/{id}/directory": {
            "get": {
                "description": "返回小说的目录",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "返回小说的目录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/book.GetBookDirectoryResp"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "测试连接是否正常",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "监控"
                ],
                "summary": "测试连接是否正常",
                "operationId": "ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ping.PingResp"
                        }
                    }
                }
            }
        },
        "/search": {
            "post": {
                "description": "返回小说的搜索结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "搜索"
                ],
                "summary": "返回小说的搜索结果",
                "operationId": "search",
                "parameters": [
                    {
                        "description": "搜索参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/search.SearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/search.SearchResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "book.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string"
                },
                "book_type": {
                    "description": "类型",
                    "type": "string"
                },
                "brief": {
                    "description": "简介",
                    "type": "string"
                },
                "img_url": {
                    "description": "图片链接",
                    "type": "string"
                },
                "last_chapter": {
                    "description": "最近更新章节",
                    "$ref": "#/definitions/book.Chapter"
                },
                "last_update_time": {
                    "description": "最近更新时间",
                    "type": "string"
                },
                "name": {
                    "description": "书名",
                    "type": "string"
                },
                "source": {
                    "description": "来源",
                    "type": "string"
                },
                "url": {
                    "description": "链接",
                    "type": "string"
                }
            }
        },
        "book.Chapter": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "章节名称",
                    "type": "string"
                },
                "url": {
                    "description": "链接",
                    "type": "string"
                }
            }
        },
        "book.GetBookDirectoryResp": {
            "type": "object",
            "properties": {
                "data_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Chapter"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "deleteat.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "model.Book": {
            "type": "object",
            "properties": {
                "book": {
                    "$ref": "#/definitions/book.Book"
                },
                "chapter_list": {
                    "description": "定义 Has Many 关系",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Chapter"
                    }
                },
                "create_at": {
                    "type": "string"
                },
                "delete_at": {
                    "$ref": "#/definitions/deleteat.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
        "model.Chapter": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "chapter": {
                    "$ref": "#/definitions/book.Chapter"
                },
                "content": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "delete_at": {
                    "$ref": "#/definitions/deleteat.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
        "ping.PingResp": {
            "type": "object",
            "properties": {
                "hello": {
                    "type": "string"
                }
            }
        },
        "search.SearchReq": {
            "type": "object",
            "required": [
                "keyword"
            ],
            "properties": {
                "keyword": {
                    "description": "搜索关键字",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "limit": {
                    "description": "限制结果数量",
                    "type": "integer",
                    "default": 10,
                    "maximum": 100,
                    "minimum": 1
                }
            }
        },
        "search.SearchResp": {
            "type": "object",
            "properties": {
                "data_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Book"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}